Changes in tst_xxx to pass compilation with new compiler
gcc9 does not recognize __rdtsc() without '#include <intrin.h>'
Also added a workaround for bug in code generator with -O3
+fixes for two minor bugs of my own that were not affecting test coverage or correctness.

+++ 128/tst_cnv_to_f.cpp
@@ -3,6 +3,7 @@
 #include <cmath>
 #include <random>
 #include <functional>           // for std::bind
+#include <intrin.h>
 #include <iostream>
 #include <boost/multiprecision/cpp_bin_float.hpp>
 
@@ -358,7 +359,7 @@ static bool test_x(extfloat128_t a, int64_t n, float ares)
     if (dNext <= dCurr) {
       bool fail = true;
       if (dNext == dCurr) {
-        uint64_t u;
+        uint32_t u;
         memcpy(&u, &ares, sizeof(u));
         fail = (u & 1) != 0;
       }
